+++ socOverviewSankey-2.1.0/release_notes.md
@@ -4,3 +4,8 @@
     - **Record Containing JSON Data** - To render all records in a single keystore record that contains a JSON object. 
     - **Get Live Data** - To dynamically render nodes and links data. The live data is queried based on the selected module and fields.
 - Added internationalization support.
+- API call updated for every module selected in each layer
+- Widget name updated
+- Minor bug fixes
+- Null values excluded while rendering Sankey chart
+- Support added for text fields in target nodes selection